fix: update tests for required invoiceRequired field and extract useControlCenterData hook
- Fix 3 failing test suites by adding invoiceRequired to test data
- Extract data-fetching logic from ControlCenterPage into useControlCenterData hook
- Reduce KYC filter limit from 10,000 to 500
